tkinter/requirements.txt
simply-sqlite
Pillow